fix: merge module
