+++ packages/core/src/internal.ts
@@ -17,7 +17,7 @@ export {
 export type { InternalContext } from './types';
 export { setHTMLPlugin, getHTMLPlugin } from './pluginHelper';
 export { formatStats, getStatsOptions, prettyTime } from './helpers';
-export { registerBuildHook, registerDevHook, onCompileDone } from './hooks';
+export { registerBuildHook, registerDevHook, onCompileDone } from './initHooks';
 export { getChainUtils, getConfigUtils } from './provider/rspackConfig';
 export { chainToConfig, modifyBundlerChain } from './configChain';
 export { applySwcDecoratorConfig } from './plugins/swc';
